+++ backend/canisters/user_index/impl/src/lifecycle/init.rs
@@ -1,7 +1,8 @@
-use crate::lifecycle::init_state;
+use crate::lifecycle::{init_state, reseed_rng};
 use crate::Data;
 use canister_tracing_macros::trace;
 use ic_cdk_macros::init;
+use std::time::Duration;
 use tracing::info;
 use user_index_canister::init::Args;
 use utils::cycles::init_cycles_dispenser_client;
@@ -13,7 +14,7 @@ fn init(args: Args) {
     canister_logger::init(args.test_mode);
     init_cycles_dispenser_client(args.cycles_dispenser_canister_id);
 
-    let env = Box::new(CanisterEnv::new());
+    let env = Box::<CanisterEnv>::default();
 
     let data = Data::new(
         args.service_principals,
@@ -30,6 +31,8 @@ fn init(args: Args) {
     );
 
     init_state(env, data, args.wasm_version);
+
+    ic_cdk::timer::set_timer(Duration::default(), reseed_rng);
 
     info!(version = %args.wasm_version, "Initialization complete");
 }