+++ backend/canisters/user_index/impl/src/lifecycle/mod.rs
@@ -1,5 +1,8 @@
-use crate::{Data, RuntimeState, WASM_VERSION};
+use crate::{mutate_state, Data, RuntimeState, WASM_VERSION};
+use tracing::trace;
 use types::{Timestamped, Version};
+use utils::canister::get_random_seed;
+use utils::env::canister::CanisterEnv;
 use utils::env::Environment;
 
 mod heartbeat;
@@ -16,4 +19,14 @@ fn init_state(env: Box<dyn Environment>, data: Data, wasm_version: Version) {
 
     crate::init_state(runtime_state);
     WASM_VERSION.with(|v| *v.borrow_mut() = Timestamped::new(wasm_version, now));
+}
+
+fn reseed_rng() {
+    ic_cdk::spawn(reseed_rng_inner());
+
+    async fn reseed_rng_inner() {
+        let seed = get_random_seed().await;
+        mutate_state(|state| state.env = Box::new(CanisterEnv::new(seed)));
+        trace!("Successfully reseeded rng");
+    }
 }