+++ backend/canisters/user_index/impl/src/model/challenges.rs
@@ -1,8 +1,8 @@
 use captcha::filters::Wave;
-use rand_chacha::rand_core::{RngCore, SeedableRng};
+use rand_core::RngCore;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use types::{Challenge, ChallengeAttempt, ChallengeKey, Salt, TimestampMillis};
+use types::{Challenge, ChallengeAttempt, ChallengeKey, TimestampMillis};
 
 const CAPTCHA_CHALLENGE_LIFETIME_MILLIS: u64 = 5 * 60 * 1000; // 5 minutes
 const MAX_INFLIGHT_CHALLENGES: u32 = 500;
@@ -11,8 +11,6 @@ const MAX_INFLIGHT_CHALLENGES: u32 = 500;
 pub struct Challenges {
     inflight: HashMap<ChallengeKey, ChallengeSolution>,
     test_mode: bool,
-    #[serde(skip)]
-    rng: Option<rand_chacha::ChaCha20Rng>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -28,19 +26,10 @@ impl Challenges {
         Challenges {
             inflight: HashMap::default(),
             test_mode,
-            rng: None,
         }
     }
 
-    pub fn is_initialised(&self) -> bool {
-        self.rng.is_some()
-    }
-
-    pub fn initialise(&mut self, seed: Salt) {
-        self.rng = Some(rand_chacha::ChaCha20Rng::from_seed(seed));
-    }
-
-    pub fn create(&mut self, now: TimestampMillis) -> Option<Challenge> {
+    pub fn create<R: RngCore>(&mut self, now: TimestampMillis, mut rng: R) -> Option<Challenge> {
         // Remove any expired challenges
         self.inflight.retain(|_, s| !s.expired(now));
 
@@ -48,42 +37,31 @@ impl Challenges {
             return None;
         }
 
-        if let Some(rng) = &mut self.rng {
-            // Try generating a unique key or trap after 10 attempts
-            const MAX_TRIES: u8 = 10;
-            for _ in 0..MAX_TRIES {
-                let key = rng.next_u32();
+        // Try generating a unique key or trap after 10 attempts
+        const MAX_TRIES: u8 = 10;
+        for _ in 0..MAX_TRIES {
+            let key = rng.next_u32();
 
-                if let std::collections::hash_map::Entry::Vacant(e) = self.inflight.entry(key) {
-                    // Create the CAPTCHA
-                    let (Base64(png_base64), chars) = create_captcha(rng);
+            if let std::collections::hash_map::Entry::Vacant(e) = self.inflight.entry(key) {
+                // Create the CAPTCHA
+                let (Base64(png_base64), chars) = create_captcha(rng);
 
-                    // Remember the solution
-                    e.insert(ChallengeSolution { created: now, chars });
+                // Remember the solution
+                e.insert(ChallengeSolution { created: now, chars });
 
-                    // Return the challenge
-                    return Some(Challenge { png_base64, key });
-                }
+                // Return the challenge
+                return Some(Challenge { png_base64, key });
             }
-
-            ic_cdk::trap(&format!("Could not find a new challenge key after {MAX_TRIES} tries"));
-        } else {
-            ic_cdk::trap("RNG not initialised");
         }
+
+        ic_cdk::trap(&format!("Could not find a new challenge key after {MAX_TRIES} tries"));
     }
 
     pub fn check(&mut self, attempt: &ChallengeAttempt, now: TimestampMillis) -> bool {
-        if self.test_mode && attempt.key == 0 && attempt.chars == "TEST" {
+        if let Some(solution) = self.inflight.remove(&attempt.key) {
+            !solution.expired(now) && solution.chars == attempt.chars
+        } else if self.test_mode && attempt.key == 0 && attempt.chars == "TEST" {
             return true;
-        }
-
-        if let Some(solution) = self.inflight.get(&attempt.key) {
-            let success = !solution.expired(now) && solution.chars == attempt.chars;
-            if !success {
-                // Remove the failed challenge so it can't be retried
-                self.inflight.remove(&attempt.key);
-            }
-            success
         } else {
             false
         }
@@ -100,7 +78,7 @@ impl ChallengeSolution {
     }
 }
 
-fn create_captcha<T: RngCore>(rng: T) -> (Base64, String) {
+fn create_captcha<R: RngCore>(rng: R) -> (Base64, String) {
     let mut captcha = captcha::RngCaptcha::from_rng(rng);
     let captcha = captcha
         .add_chars(5)