+++ backend/libraries/utils/src/env.rs
@@ -1,4 +1,5 @@
 use candid::Principal;
+use rand::rngs::StdRng;
 use types::{CanisterId, Cycles, TimestampMillis, TimestampNanos};
 
 pub mod canister;
@@ -8,9 +9,8 @@ pub trait Environment {
     fn now_nanos(&self) -> TimestampNanos;
     fn caller(&self) -> Principal;
     fn canister_id(&self) -> CanisterId;
-    fn random_u32(&mut self) -> u32;
-    fn random(&mut self) -> f64;
     fn cycles_balance(&self) -> Cycles;
+    fn rng(&mut self) -> &mut StdRng;
 
     fn now(&self) -> TimestampMillis {
         self.now_nanos() / 1_000_000